% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean the occurrence records}
\usage{
clean(coord, abio)
}
\arguments{
\item{coord}{data.frame with the coordinates}

\item{abio}{an object with the rasters to be cut. Accepts an object of type RasterStack, generated by the function \code{\link[raster]{stack}}}
}
\value{
data.frame containing longitude and latitude.
}
\description{
A function to remove the points that are outside the extension of the raster and to maintain, at most one register per pixel.
}
\details{
Used internally in function \code{\link[modelos]{modelos}}.
}
\examples{
fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''), pattern='grd', full.names=TRUE )
predictors <- raster::stack(fnames)
clean(coord = manimax, abio = predictors)

}
\seealso{
\code{\link[raster]{cellFromXY}}, \code{\link[raster]{mask}}, \code{\link[raster]{extract}}
}
\author{
Diogo S. B. Rocha
}
