% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_data.R
\name{cor.data}
\alias{cor.data}
\title{Exibe um gráfico com as correlações entre as variáveis}
\usage{
cor.data(abio, plot = TRUE, method = "pearson", rep = 1000)
}
\arguments{
\item{abio}{os rasters. Objeto do tipo _stack_}

\item{plot}{lógico. Plota um dos rasters cortados.}

\item{method}{a character string indicating which correlation coefficient (or covariance) is to be computed: "pearson" (default), "kendall", or "spearman".}

\item{rep}{numérico. Número de pontos gerados para extrair os valores dos raters e utilizar na correlação. O padrão é 1000.}
}
\value{
Retorna uma tabela com os valores de correlação entre as variáveis.
}
\description{
Uma funcao para exibir um gráfico com as correla??es entre as vari?veis ambientais informadas.
}
\details{
O índice de correlação utilizado é spearman
}
\examples{
fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''), pattern='grd', full.names=TRUE )
predictors <- stack(fnames)
cor.data(abio = predictors)

}
\seealso{
\code{\link[graphics]{pairs}}
}
\author{
Diogo S. B. Rocha
}
