% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_raster.R
\name{cut.raster}
\alias{cut.raster}
\title{Cut raster based on a shape}
\usage{
cut.raster(abio, shape.dir, extension = ".asc", br = FALSE, plot = TRUE,
  diretorio = "cortados", trim = FALSE)
}
\arguments{
\item{abio}{an object with the rasters to be cut. Accepts an object of type RasterStack, generated by the function \code{\link[raster]{stack}}}

\item{shape.dir}{here are two options. The first is to inform an object in the shapefile format. The second option is to inform a directory containing the shapefile.}

\item{extension}{extent of the cut rasters. Default is .asc, see\code{\link[raster]{write.raster}} for more output format possibilities.}

\item{br}{logical. If TRUE, use the Brazil shape provided by \code{\link[maptools]{wrld_simpl}}. If FALSE (default), use the informed shape.}

\item{plot}{logical. Plots one of the cut rasters.}

\item{trim}{logical. If TRUE, exclude pixels with NA (very slow). If FALSE (default), keep the NAs.}
}
\value{
Raster files in a folder "Cortados".
}
\description{
A function to cut a raster based on an informed shapefile.
}
\details{
When the shapeir argument is a folder where it contains the reference shape, this folder must contain only the shape and its auxiliary files.
}
\examples{

fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''), pattern='grd', full.names=TRUE )
predictors <- raster::stack(fnames)
cut.raster(abio = predictors, br=T, extension = ".tif")

}
\seealso{
\code{\link[raster]{crop}}, \code{\link[raster]{mask}}, \code{\link[raster]{stack}}
}
\author{
Diogo S. B. Rocha
}
