% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_filt.R
\name{geo.filt}
\alias{geo.filt}
\title{Geographic filter}
\usage{
geo.filt(coord, resolution = 10)
}
\arguments{
\item{coord}{data.frame. Table with the species' data. It should contain only two columns: long and lat, in that order.}

\item{res}{numeric. Minimum distance (in Km) between points.}
}
\value{
Table of occurrence points with minimum distance indicated in res
}
\description{
A function to generate ecological niche models.
}
\details{
Nothing for now.
}
\examples{
geo.filt(pts=manimax,res=20)

}
\seealso{
\code{\link[dismo]{gridSample}}
}
\author{
Diogo S. B. Rocha
}
