% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_mx.R
\name{i.mx}
\alias{i.mx}
\title{Download the file maxen.jar and put it in the folder of the dismo}
\usage{
i.mx()
}
\arguments{
\item{no_params}{No arguments}
}
\value{
Arquivos raster em um diretório indicado pelo usuário.
}
\description{
A function to automate the download of maxent.jar file
}
\details{
Just download the maxent.jar file from the site: \link[http://biodiversityinformatics.amnh.org/open_source/maxent/maxent.php?op=download ] and places in the package folder dismo
}
\examples{
i.mx()

}
\seealso{
\code{\link[dismo]{bioclim}}, \code{\link[raster]{mask}}
}
\author{
Diogo S. B. Rocha
}
