% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelos4.R
\name{modelos}
\alias{modelos}
\title{Gera os modelos}
\usage{
modelos(coord, abio, k = 3, diretorio = "teste", plot = T, bc = T,
  mx = F, GLM = F, RF = F, SVM = F, dm = F, mah = F, proj,
  buffer = "none", geo.filt.res, mod = "before", tss = 0)
}
\arguments{
\item{coord}{data.frame. Tabela com os dados de ocorrencia da espécie. Deve conter apenas duas colunas: long e lat, nesta ordem.}

\item{abio}{os rasters a serem cortados. Aceita um objeto do tipo _stack_}

\item{k}{número de partições. O padrão são 3.}

\item{diretorio}{nome do diretório a ser criado com os resultados na modelagem.}

\item{plot}{lógico. Plota o modelo final.}

\item{bc}{bioclim}

\item{mx}{Maxent}

\item{GLM}{Generalized linear model}

\item{RF}{Random Forest (regression)}

\item{SVM}{Support Vector Machine}

\item{dm}{Domain}

\item{mah}{distância de Mahalanobis}

\item{proj}{stack com as variáveis onde será projetado o modelo. Caso nao seja informado, o modelo é projetado no mesmo local de cria??o do modelo (informado em abio).}

\item{buffer}{distância escolhida para gerar um buffer em torno dos pontos de ocorrência onde será gerados os pontos de pseudo-aus?ncia. "mean" é a distância média entre os pontos e "max" é a distância máxima entre os pontos. Ou se for "none", não é usado nehum buffer (padrão).}

\item{geo.filt.res}{numerico. Mantem apenas os pontos que estejam, no mínimo, distantes um do outro o numero de km informado.}

\item{mod}{quando o modelo é cortado para gerar o ensemble. "before" cada partição é cortada pelo seu próprio TSSth. "after" o ensemble de cada algoritmo é cortado pelo TSSth médio das partições.}

\item{tss}{numérico. Seleciona apenas modelos que apresente valor TSS maior do que o informado.}
}
\value{
Arquivos raster em um diretório indicado pelo usuário.
}
\description{
Uma funcao para geral modelos de nicho ecológico.
}
\details{
A função mais complexa deste pacote
}
\examples{
fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''), pattern='grd', full.names=TRUE )
predictors <- raster::stack(fnames)
occurence <- paste(system.file(package="dismo"), '/ex/bradypus.csv', sep='')
occ <- read.table(occurence, header=TRUE, sep=',')[,-1]
modelos(coord = occ, abio = predictors)

}
\seealso{
\code{\link[dismo]{bioclim}}, \code{\link[raster]{mask}}
}
\author{
Diogo S. B. Rocha
}
