% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Limpas os registros de ocorrência}
\usage{
clean(coord, abio)
}
\arguments{
\item{coord}{data.frame com as coordenadas}

\item{abio}{os rasters a serem cortados. Aceita um objeto do tipo _stack_}
}
\value{
data.frame contendo longitude e latitude.
}
\description{
Uma funcao para retirar os pontos que est?o fora da extens?o do raster e manter, no m?ximo um registro por pixel.
}
\details{
aaa
}
\examples{
require(dismo)
fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''), pattern='grd', full.names=TRUE )
predictors <- stack(fnames)
cut.raster(abio = predictors, br=T, extension = ".tif")

}
\seealso{
\code{\link[raster]{cellFromXY}}, \code{\link[raster]{mask}}, \code{\link[raster]{extract}}
}
\author{
Diogo S. B. Rocha
}
