% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Limpas os registros de ocorrência}
\usage{
clean(coord, abio)
}
\arguments{
\item{coord}{data.frame com as coordenadas}

\item{abio}{um objeto com os rasters a serem cortados. Aceita um objeto do tipo RasterStack, gerado pela função \code{\link[raster]{stack}}}
}
\value{
data.frame contendo longitude e latitude.
}
\description{
Uma função para retirar os pontos que estão fora da extensão do raster e manter, no máximo um registro por pixel.
}
\details{
Esta função é usada internamente na função \code{\link[modelos]{modelos}}.
}
\examples{
fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''), pattern='grd', full.names=TRUE )
predictors <- raster::stack(fnames)
clean(coord = manimax, abio = predictors)

}
\seealso{
\code{\link[raster]{cellFromXY}}, \code{\link[raster]{mask}}, \code{\link[raster]{extract}}
}
\author{
Diogo S. B. Rocha
}
