% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_raster.R, R/modelos4.R
\name{cut.raster}
\alias{cut.raster}
\alias{cut.raster}
\alias{modelos}
\title{Corta raster com base em um shape}
\usage{
cut.raster(abio, shape.dir, extension = ".asc", br = FALSE, plot = TRUE,
  trim = FALSE)

modelos(coord, abio, k = 3, diretorio = "teste", plot = T, bc = T,
  mx = F, GLM = F, RF = F, SVM = F, dm = F, mah = F, proj,
  buffer = "none", geo.filt = T, mod = "before", tss = 0)
}
\arguments{
\item{abio}{os rasters a serem cortados. Aceita um objeto do tipo _stack_}

\item{shape.dir}{aqui tem duas op??es. A primeira ? informar um ojeto no formato shapefile. A segunda op??o ? informar um diret?rio que contem o shapefile.}

\item{extension}{extens?o de sa?da dos rasters cortados. O padr?o ? .asc, veja \code{\link[raster]{write.raster}} para mais possibilidades de formatos de sa?da.}

\item{br}{l?gico. Se TRUE, utiliza o shape do brasil formecido por \code{\link[maptools]{wrld_simpl}}. Se FALSE (padr?o), utliza o shape informado.}

\item{plot}{l?gico. Plota um dos rasters cortados.}

\item{trim}{l?gico. Se TRUE, exclui os pixels com NA (mais demorado). Se FALSE (padr?o), mant?m os NAs.}

\item{coord}{data.frame. Tabela com os dados de oco?ncia da esp?cie. Deve conter apenas duas colunas: long e lat, nesta ordem.}

\item{k}{n?mero de parti??es. Padr?o ? 3.}

\item{diretorio}{nome do diret?rio a ser criado com os resultados na modelagem.}

\item{bc}{bioclim}

\item{mx}{Maxent}

\item{GLM}{Generalized linear model}

\item{RF}{Random Forest (regression)}

\item{SVM}{Support Vector Machine}

\item{dm}{Domain}

\item{mah}{dist?ncia de Mahalanobis}

\item{proj}{stack com as vari?veis onde ser? projetado o modelo. Caso n?o seja informado, o modelo ? projetado no mesmo local de cria??o do modelo (informado em abio).}

\item{buffer}{dist?ncia escolhida para gerar um buffer em torno dos pontos de ocorr?ncia onde ser? gerados os pontos de pseudo-aus?ncia. "mean" ? a dist?ncia m?dia entre os pontos e "max" ? a dist?ncia m?xima entre os pontos.}

\item{geo.filt}{l?gico. Se TRUE (padr?o), mant?m apenas os pontos que estejam no m?nimo 20km distantes um do outro.}

\item{mod}{quando o modelo ? cortado para gerar o ensemble. "before" cada parti??o ? cortada pelo seu pr?prio TSSth. "after" o ensemble de cada algoritmo ? cortado pelo TSSthm?dio das parti??es.}

\item{tss}{num?rico. Seleciona apenas modelos que apresente valor TSS maior do que o informado.}

\item{abio}{os rasters a serem cortados. Aceita um objeto do tipo _stack_}

\item{plot}{l?gico. Plota o modelo final.}
}
\value{
Arquivos raster em uma pasta chamada "Cortados".

Arquivos raster em um diret?rio indicado pelo usu?rio.
}
\description{
Uma funcao para cortar um raster com base em um shapefile informado.

Uma funcao para cortar um raster com base em um shapefile informado.
}
\details{
Quando argumento shapedir for uma pasta onde cont?m o shape de refer?ncia, esta pasta deve conter somente o shape e seus arquivos auxiliares.

A fun??o mais complexa deste pacote
}
\examples{
fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''), pattern='grd', full.names=TRUE )
predictors <- stack(fnames)
cut.raster(abio = predictors, br=T, extension = ".tif")

fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''), pattern='grd', full.names=TRUE )
predictors <- stack(fnames)
occurence <- paste(system.file(package="dismo"), '/ex/bradypus.csv', sep='')
occ <- read.table(occurence, header=TRUE, sep=',')[,-1]
modelos(coord = occ, abio = predictors)

}
\seealso{
\code{\link[raster]{crop}}, \code{\link[raster]{mask}}

\code{\link[dismo]{bioclim}}, \code{\link[raster]{mask}}
}
\author{
Diogo S. B. Rocha

Diogo S. B. Rocha
}
