% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_raster.R
\name{cut.raster}
\alias{cut.raster}
\title{Corta raster com base em um shape}
\usage{
cut.raster(abio, shape.dir, extension = ".asc", br = FALSE, plot = TRUE,
  trim = FALSE)
}
\arguments{
\item{abio}{os rasters a serem cortados. Aceita um objeto do tipo _stack_}

\item{shape.dir}{aqui tem duas opções. A primeira é informar um ojeto no formato shapefile. A segunda opção é informar um diretório que contem o shapefile.}

\item{extension}{extensão de saída dos rasters cortados. O padrão é .asc, veja \code{\link[raster]{write.raster}} para mais possibilidades de formatos de sa?da.}

\item{br}{lógico. Se TRUE, utiliza o shape do brasil formecido por \code{\link[maptools]{wrld_simpl}}. Se FALSE (padrão), utliza o shape informado.}

\item{plot}{lógico. Plota um dos rasters cortados.}

\item{trim}{lógico. Se TRUE, exclui os pixels com NA (mais demorado). Se FALSE (padr?o), mant?m os NAs.}
}
\value{
Arquivos raster em uma pasta chamada "Cortados".
}
\description{
Uma funcao para cortar um raster com base em um shapefile informado.
}
\details{
Quando argumento shapedir for uma pasta onde contém o shape de referência, esta pasta deve conter somente o shape e seus arquivos auxiliares.
}
\examples{
fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''), pattern='grd', full.names=TRUE )
predictors <- stack(fnames)
cut.raster(abio = predictors, br=T, extension = ".tif")

}
\seealso{
\code{\link[raster]{crop}}, \code{\link[raster]{mask}}
}
\author{
Diogo S. B. Rocha
}
