% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_filt.R
\name{geo.filt}
\alias{geo.filt}
\title{Filtro geográfico}
\usage{
geo.filt(coord, resolution = 10)
}
\arguments{
\item{coord}{data.frame. Tabela com os dados de oco?ncia da espécie. Deve conter apenas duas colunas: long e lat, nesta ordem.}

\item{res}{numerico. distância mínima (em Km) entre os pontos.}
}
\value{
tabela de pontos de ocorrência com distância mínima indicada em res
}
\description{
Uma funcao para geral modelos de nicho ecológico.
}
\details{
Nada por enquanto
}
\examples{
geo.filt(pts=manimax,res=20)

}
\seealso{
\code{\link[dismo]{gridSample}}
}
\author{
Diogo S. B. Rocha
}
