% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelos4.R
\name{modelos}
\alias{modelos}
\title{Produces ecological niche models}
\usage{
modelos(coord, abio, k = 3, diretorio = "teste", plot = T, bc = T,
  mx = F, GLM = F, RF = F, SVM = F, dm = F, mah = F, proj,
  buffer = "none", geo.filt.res, mod = "before", tss = 0)
}
\arguments{
\item{coord}{data.frame. Table with the occurrence data of the species. It should contain only two columns: long and lat, in this order.}

\item{abio}{the rasters to be cut. Accepts an object of the type _stack_}

\item{k}{number of partitions. The default is 3.}

\item{diretorio}{name of the directory to be created with the results in the modeling.}

\item{plot}{logical. If TRUE (deafult), plots the final model.}

\item{bc}{bioclim}

\item{mx}{Maxent}

\item{GLM}{Generalized linear model}

\item{RF}{Random Forest (regression)}

\item{SVM}{Support Vector Machine}

\item{dm}{Domain}

\item{mah}{Mahalanobis distance}

\item{proj}{_stack_ with the variables where the model will be projected. If you are not informed, the model is projected in the same place of creation of the model (informed in abio).}

\item{buffer}{distance chosen to generate a buffer around the occurrence points where the pseudo-absence points will be generated. "mean" is the mean distance between points, "median" is the median distance between points and "max" is the maximum distance between points. Or if it is "none", no buffer is used (default).}

\item{geo.filt.res}{numeric. Keep only the points that are at least far from each other the number of kilometers informed.}

\item{mod}{when the model is cut to generate the ensemble. "before" each partition is cut by its own TSSth. "after" the ensemble of each algorithm is cut by the average TSSth of the partitions.}

\item{tss}{numeric. Selects only models that present higher than reported TSS value.}
}
\value{
Raster files in a user-specified directory.
}
\description{
A function to produce ecological niche models.
}
\details{
The most complex function of this package
}
\examples{
fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''), pattern='grd', full.names=TRUE )
predictors <- raster::stack(fnames)
occurence <- paste(system.file(package="dismo"), '/ex/bradypus.csv', sep='')
occ <- read.table(occurence, header=TRUE, sep=',')[,-1]
modelos(coord = occ, abio = predictors)

}
\seealso{
\code{\link[dismo]{bioclim}}, \code{\link[raster]{mask}}
}
\author{
Diogo S. B. Rocha
}
