% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toKML.R
\name{toKML}
\alias{toKML}
\title{Transforma o raster em arquivo KML}
\usage{
toKML(modelo, name = "meuKML", zeros = FALSE, open = FALSE)
}
\arguments{
\item{modelo}{raster com o modelo de distribui??o}

\item{name}{nome do arquivo KML gerado.}

\item{zeros}{lógico. Se FALSE, os zeros do modelo s?o convertivos em NA.}

\item{open}{lógico. Se TURE, abre o arquivo KML no Google Earth.}
}
\value{
data.frame contendo longitude e latitude.
}
\description{
Uma função para transformar o modelo de nicho de formato raster em arquivo KML.
}
\details{
This function is a wrapper for \code{\link[plotKML]{plotKML}} the function from the roxygen2 package. See the documentation and vignettes of that package to learn how to use roxygen.
}
\examples{
fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''), pattern='grd', full.names=TRUE )
predictors <- stack(fnames)
toKML(modelo = predictors, name = "exemplo")


}
\seealso{
\code{\link[plotKML]{plotKML}}
}
\author{
Diogo S. B. Rocha
}
